Move services into their own package
